test: seed database via shared fixture
