+++ core/tests/__init__.py
@@ -1,5 +1,1 @@
-from .test_admin_views import *
-from .test_forms import *
-from .test_parsing import *
-from .test_general import *
-from .test_gap_notes import *
+"""Testpaket für die core-Anwendung."""
